requirements.txt
altair==5.5.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.3
curl_cffi==0.12.0
cycler==0.12.1
eventkit==1.0.3
fonttools==4.59.0
frozendict==2.4.6
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.3
ib-insync==0.9.86
idna==3.10
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.5
multitasking==0.0.12
narwhals==2.0.1
nest-asyncio==1.6.0
numpy==2.3.2
packaging==25.0
pandas==2.3.1
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.8
plotly==6.6.0
protobuf==6.31.1
psycopg2-binary==2.9.10
pyarrow==21.0.0
pycparser==2.22
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
six==1.17.0
smmap==5.0.2
soupsieve==2.7
SQLAlchemy==2.0.42
streamlit==1.48.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
yfinance==0.2.65
